% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_window.R
\name{generic_window_check}
\alias{generic_window_check}
\title{Generic checks for window constructors}
\usage{
generic_window_check(window_dt)
}
\arguments{
\item{window_dt}{A data.table}
}
\value{
TRUE if the checks pass, otherwise an error is raised.
}
\description{
Runs the checks every window constructor needs before it builds windows.
}
\keyword{internal}
