% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_selector.R
\name{selector_object_name}
\alias{selector_object_name}
\title{Name a Custom Selector Would Be Looked Up Under}
\usage{
selector_object_name(selector_name)
}
\arguments{
\item{selector_name}{Value of the metadata \code{selector} column.}
}
\value{
The \verb{<selector>_selector} object name, lower-cased.
}
\description{
Name a Custom Selector Would Be Looked Up Under
}
\keyword{internal}
