% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_registry.R
\name{selector_sql_path}
\alias{selector_sql_path}
\title{Get the Path to a Selector SQL Template}
\usage{
selector_sql_path(selector, package = "anchoR")
}
\arguments{
\item{selector}{Selector name such as \code{"LATEST"} or \code{"COUNT"}.}

\item{package}{Package name used to resolve \code{inst/sql}.}
}
\value{
An absolute path to the SQL template.
}
\description{
Get the Path to a Selector SQL Template
}
