% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_registry.R
\name{available_selectors}
\alias{available_selectors}
\title{List Available Selector SQL Templates}
\usage{
available_selectors(package = "anchoR")
}
\arguments{
\item{package}{Package name used to resolve \code{inst/sql}.}
}
\value{
A character vector of selector names.
}
\description{
List Available Selector SQL Templates
}
